Minor fixes.

+++ model/rbac_test.go
@@ -80,7 +80,7 @@ func TestRBAC(t *testing.T) {
 		if err != nil {
 			t.Error(err)
 		} else {
-			assert.Equal(t, uint8(1), roles[0].Type)
+			assert.Equal(t, uint(1), roles[0].Type)
 		}
 
 		if err := tRepo.mock.ExpectationsWereMet(); err != nil {
